moving files to lab2_var3
